backend/requirements.txt
accelerate
aiohttp
aiosignal
attrs
blinker
certifi
charset-normalizer
click
colorama
coloredlogs
contourpy
cycler
datasets
dill
et-xmlfile
filelock
Flask
Flask-Cors
fonttools
frozenlist
fsspec
huggingface-hub
humanfriendly
idna
intel-openmp
itsdangerous
Jinja2
joblib
kiwisolver
MarkupSafe
matplotlib
mkl
mpmath
multidict
multiprocess
networkx
numpy
openpyxl
optimum
packaging
pandas
pillow
protobuf
psutil
pyarrow
pyarrow-hotfix
pyparsing
pyreadline3
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentencepiece
six
sympy
tbb
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers
typing_extensions
tzdata
urllib3
Werkzeug
xxhash
yarl
